+++ nominal/core/client.py
@@ -18,6 +18,7 @@ from nominal_api import (
     event,
     ingest_api,
     scout_asset_api,
+    scout_catalog,
     scout_checks_api,
     scout_datasource_connection_api,
     scout_notebook_api,
@@ -38,6 +39,7 @@ from nominal.core._utils import (
     construct_user_agent_string,
     create_search_assets_query,
     create_search_checklists_query,
+    create_search_datasets_query,
     create_search_events_query,
     create_search_runs_query,
     create_search_secrets_query,
@@ -48,6 +50,7 @@ from nominal.core._utils import (
     search_assets_paginated,
     search_checklists_paginated,
     search_data_reviews_paginated,
+    search_datasets_paginated,
     search_events_paginated,
     search_runs_paginated,
     search_secrets_paginated,
@@ -239,6 +242,39 @@ class NominalClient:
         """
         query = create_search_users_query(exact_match, search_text)
         return list(self._iter_search_users(query))
+
+    def _iter_search_datasets(self, query: scout_catalog.SearchDatasetsQuery) -> Iterable[Dataset]:
+        for raw_dataset in search_datasets_paginated(self._clients.catalog, self._clients.auth_header, query):
+            yield Dataset._from_conjure(self._clients, raw_dataset)
+
+    def search_datasets(
+        self,
+        *,
+        exact_match: str | None = None,
+        search_text: str | None = None,
+        labels: Sequence[str] | None = None,
+        properties: Mapping[str, str] | None = None,
+        before: str | datetime | IntegralNanosecondsUTC | None = None,
+        after: str | datetime | IntegralNanosecondsUTC | None = None,
+        workspace_rid: str | None = None,
+    ) -> Sequence[Dataset]:
+        """Search for datasets the specified filters.
+        Filters are ANDed together, e.g. `(secret.label == label) AND (secret.property == property)`
+
+        Args:
+            exact_match: Searches for an exact substring of dataset name
+            search_text: Searches for a (case-insensitive) substring across all text fields.
+            labels: A sequence of labels that must ALL be present on a secret to be included.
+            properties: A mapping of key-value pairs that must ALL be present on a secret to be included.
+            before: Searches for datasets created before some time (inclusive).
+            after: Searches for datasets created before after time (inclusive).
+            workspace_rid: Filters search to given workspace. If None, searches within default workspace
+
+        Returns:
+            All datasets which match all of the provided conditions
+        """
+        query = create_search_datasets_query(exact_match, search_text, labels, properties, before, after, workspace_rid)
+        return list(self._iter_search_datasets(query))
 
     def get_workspace(self, workspace_rid: str | None = None) -> Workspace:
         """Get workspace via given RID, or the default workspace if no RID is provided.